optimize: Improvement for PSReadLine check (#79)
Calls to Get-Module with -ListAvailable are potentially quite expensive.
Try to avoid it by checking if PSReadLine is already imported, and
falling back to enumerating all available modules only if it's not.

On Windows 10, PSReadLine is included with the operating system and
automatically imported, even if PowerShell is launched with -NoProfile.
It is not imported if PowerShell is launched with -NonInteractive. This
should mean PSReadLine is already imported on most Windows installs.

Testing on my own system shows this optimisation saves around ~200ms.